Change 'board name' to a function component
